+++ CHANGES.md
@@ -13,6 +13,7 @@ Release Notes.
 - Add the sub-command `process estimate scale` to estimate the process scale.(#142)
 - Update the `process list` to must be provided the instance and duration.(#144)
 - Remove `layer` field in the `instance` and `process` commands for 9.1.0 GraphQL query protocol's breaking changes (#149)
+- Remove `duration` flag in `profiling ebpf schedules`.(#150)
 
 0.10.0
 ------------------